Make map interaction helper
